Episode API

+++ app/src/main/java/com/example/videostreaming/screen/model/Content.kt
@@ -3,7 +3,7 @@ package com.example.videostreaming.screen.model
 data class Content(
     val id: String="",
     val episodeId: String="",
-    val episodeNumber: Int=0,
+    val episodeNumber: Double=0.0,
     val title: String="",
     val image: String="",
     val url: String=""
